# swagger.yaml
swagger: '2.0'
info:
  description: How to use the Boastable server.
  version: 1.0.0
  title: Boastable API
  contact:
    email: ryan@fueledonbacon.com
host: localhost:3000
basePath: /
...
schemes:
  - http
paths:
...
  /stripe/{model}/list:
    get:
      tags:
        - stripe
      summary: "Retrieve a list of stripe documents of a particular {model} type"
      description: "This is an admin permissioned endpoint. Gets lists of stripe data"
      operationId: getStripeDataList
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: path
          name: model
          description: One of the Stripe document types
          required: true
        - in: query
          name: query
          description: Any key on the membership document
          required: true
      responses:
        '200':
          description: A stripe document list
        '405':
          description: Invalid input
...
  /stripe/{model}:
    get:
      tags:
        - stripe
      summary: "Retrieve a stripe document"
      description: "This is an admin permissioned endpoint. It allows for the querying of raw stripe data using query parameters"
      operationId: getStripeData
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: path
          name: model
          description: One of the Stripe document types
          required: true
        - in: query
          name: query
          description: Any key on the membership document
          required: true
      responses:
        '200':
          description: A stripe response
        '405':
          description: Invalid input
  /user/list:
    get:
      tags:
        - user
      summary: Get a list of users
      description: 'this is an admin only function'
      operationId: listUsers
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: id
          in: query
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /user:
    post:
      tags:
        - user
      summary: Create user document
      description: 'An administrator can create any user, but a client can only create their own. A client user document is created at register'
      operationId: createUser
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
    get:
      tags:
        - user
      summary: Get user by user name
      description: 'Admin get get any user, Client can only get their own profile'
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Updated user
      description: This can only be done by the logged in user. Client updates self, Admin can update anyone.
      operationId: updateUser
      produces:
        - application/xml
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: username
          in: path
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user. Admin role only.
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: an authorization header
          example: bearer [insert JWT token here]
          required: true
          type: string
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
...
definitions:
...
  User:
    type: object
    required:
      - auth
    properties:
      _id:
        type: string
        description: Mongo document ObjectId
        example: 5a2cfe672ebedf740dac80c8
      auth:
        type: string
        description: ObjectId reference to an Auth document
        example: 5a2cfe672ebedf740dac80c8
      first_name:
        type: string
        example: Bilbro
      last_name:
        type: string
        example: Baggins
      phone:
        type: string
        description: An E.164 formatted string
        example: +15615236666
...